+++ src/test/java/kitchenpos/deliveryorder/infra/FakeKitchenridersClient.java
@@ -1,6 +1,4 @@
-package kitchenpos.application;
-
-import kitchenpos.infra.KitchenridersClient;
+package kitchenpos.deliveryorder.infra;
 
 import java.math.BigDecimal;
 import java.util.UUID;
