+++ src/test/java/kitchenpos/eatinorder/domain/InMemoryOrderTableRepository.java
@@ -1,9 +1,11 @@
-package kitchenpos.application;
+package kitchenpos.eatinorder.domain;
 
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.OrderTableRepository;
-
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
 
 public class InMemoryOrderTableRepository implements OrderTableRepository {
     private final Map<UUID, OrderTable> orderTables = new HashMap<>();
